+++ tests/test_bitmap.c
@@ -1,21 +1,11 @@
 #include "collection/bitmap.h"
 #include "memory/allocator.h"
-#include "memory/fixedallocator.h"
-#include "memory/wasmallocator.h"
+#include "test_common.h"
 #include "unity.h"
 #include <unity_internals.h>
 
 static void Bitmap_Basic(void) {
-#if CU_PLAT_WASM
-  cu_Allocator alloc = cu_Allocator_WasmAllocator();
-#elif CU_FREESTANDING
-  static char buf[1024];
-  cu_FixedAllocator fa;
-  cu_Allocator alloc =
-      cu_Allocator_FixedAllocator(&fa, cu_Slice_create(buf, sizeof(buf)));
-#else
-  cu_Allocator alloc = cu_Allocator_CAllocator();
-#endif
+  cu_Allocator alloc = test_allocator;
   cu_Bitmap_Optional opt = cu_Bitmap_create(alloc, 128);
   TEST_ASSERT_TRUE(cu_Bitmap_Optional_is_some(&opt));
   cu_Bitmap map = opt.value;